Migration Mutation.deleteAlerts
