+++ packages/services/api/src/modules/alerts/resolvers.ts
@@ -1,49 +1,8 @@
-import { HiveError } from '../../shared/errors';
-import { ProjectManager } from '../project/providers/project-manager';
-import { IdTranslator } from '../shared/providers/id-translator';
 import { TargetManager } from '../target/providers/target-manager';
 import type { AlertsModule } from './__generated__/types';
 import { AlertsManager } from './providers/alerts-manager';
 
 export const resolvers: AlertsModule.Resolvers = {
-  Mutation: {
-    async deleteAlerts(_, { input }, { injector }) {
-      const translator = injector.get(IdTranslator);
-      const [organizationId, projectId] = await Promise.all([
-        translator.translateOrganizationId(input),
-        translator.translateProjectId(input),
-      ]);
-
-      const project = await injector.get(ProjectManager).getProject({
-        organization: organizationId,
-        project: projectId,
-      });
-
-      try {
-        await injector.get(AlertsManager).deleteAlerts({
-          organization: organizationId,
-          project: projectId,
-          alerts: input.alerts,
-        });
-
-        return {
-          ok: {
-            updatedProject: project,
-          },
-        };
-      } catch (error) {
-        if (error instanceof HiveError) {
-          return {
-            error: {
-              message: error.message,
-            },
-          };
-        }
-
-        throw error;
-      }
-    },
-  },
   Project: {
     async alerts(project, _, { injector }) {
       return injector.get(AlertsManager).getAlerts({
